+++ ProjectOne/ProjectOne/src/app/app.config.ts
@@ -1,9 +1,26 @@
 import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
-import { provideRouter } from '@angular/router';
+import { provideRouter, Routes } from '@angular/router';
 import { providePrimeNG } from 'primeng/config';
 import Aura from '@primeng/themes/aura';
 import { provideAnimations } from '@angular/platform-browser/animations';
-import { routes } from './routes';
+import {InicioComponent} from './inicio/inicio.component';
+import {RegistroComponent} from './registro/registro.component';
+import {PrincipalComponent} from './principal/principal.component';
+import {UsuarioComponent} from './usuario/usuario.component';
+import {PokemonesComponent} from './pokemones/pokemones.component';
+import {EquiposComponent} from './equipos/equipos.component';
+import {MusicaComponent} from './musica/musica.component';
+
+const routes: Routes = [
+  { path: '', redirectTo: 'inicio', pathMatch: 'full' },
+  { path: 'inicio', component: InicioComponent },
+  { path: 'registro', component: RegistroComponent },
+  { path: 'principal', component: PrincipalComponent },
+  { path: 'usuario', component: UsuarioComponent },
+  { path: 'pokemones', component: PokemonesComponent },
+  { path: 'equipos', component: EquiposComponent },
+  { path: 'musica', component: MusicaComponent }
+];
 
 export const appConfig: ApplicationConfig = {
   providers: [
