+++ ProjectOne/ProjectOne/src/app/equipos/equipos.component.ts
@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-equipos',
-  standalone: false,
+  standalone: true,
   templateUrl: './equipos.component.html',
-  styleUrl: './equipos.component.css'
+  styleUrls: ['./equipos.component.css']
 })
 export class EquiposComponent {
