+++ ProjectOne/ProjectOne/src/app/pokemones/pokemones.component.css
@@ -1,0 +1,4 @@
+p{
+  font-family: Arial, Helvetica, sans-serif;
+  background: #881212;
+}
