+++ ProjectOne/ProjectOne/src/app/principal/principal.component.html
@@ -5,8 +5,7 @@
   </button>
 
   <!-- Sidebar/Drawer -->
-  <p-drawer [(visible)]="visible" position="left" [modal]="true" [dismissable]="true" (onHide)="visible = false" class="custom-sidebar">
-    <div class="sidebar-content">
+  <p-drawer [(visible)]="visible" position="left" [modal]="true" [dismissible]="true" (onHide)="visible = false" class="custom-sidebar">    <div class="sidebar-content">
       <!-- Opciones Sidebar -->
       <div class="sidebar-options">
         <div class="sidebar-title">MENÚ</div>
