+++ ProjectOne/ProjectOne/src/app/usuario/usuario.component.ts
@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-usuario',
-  standalone: false,
+  standalone: true,
   templateUrl: './usuario.component.html',
-  styleUrl: './usuario.component.css'
+  styleUrls: ['./usuario.component.css']
 })
 export class UsuarioComponent {
